refactor: remove definitions.py + merge Key/BaseKey + various improvements
